Upload of The Latest State of The Solution as The Initial Version (1.0)
Please use the solution in the following order for proper test functionality:
(1) Select a course
(2) Select a team (you have to designate the whole class as a single team for proper functionality. You may select teams 99 OR 999 for this purpose)
(3) Evaluate!

Use the plain tables to open new courses, add students or input evaluation criterion.